pyproject.toml
[tool.poetry]
name = "orca_backend"
version = "0.1.0"
description = "Orca backend web server implemets orca_nw_lib"
readme = "readme.md"
authors = ["kamal bhatt <kamal.bhatt@stordis.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.8"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
pytest = "^7.3.2"
pytest-django = "^4.7.0"
orca-nw-lib = "^1.3.3"
django-oauth-toolkit="^2.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
